Add lib/localization_lqsr.py
    * move code lsqr_localization.py -> lib/localization_lqsr.py
    * add new class localization_lqsr.LocalizationByLeastSquares
